Add CardGeneratorImpl and CardGenerator Interface
This commit introduces CardGeneratorImpl, a new class that implements CardGenerator interface. It includes the logic to generate new game cards with random attributes such as name, description, imageUrl, family, affinity, hp, energy, attack, defense and price. The CardGenerator interface, also added in this commit, outlines the method 'generateNewCard', which is fully implemented in the CardGeneratorImpl.