Add unit tests for CardController and AuthController
Tests have been added for the CardController and the AuthController classes to verify the functionality of the buyCardsToMarket method and the register method respectively. These tests make use of Mockito to mock the AuthService, CardService, and MarketService classes, as well as the required HTTP request objects. A separate configuration file for test environments, named application-test.properties has also been added.