Refactor and enhance handling of card, money, and user operations
This commit refactors the handling of operations associated with cards, money, and user. It includes renaming of the `CardCheck` class to `CardBuyable` and deletion of the `bpm-plateform.xml` file. It also handles access failures to the platform more robustly by catching RestClientExceptions and throwing BpmnErrors in response. New classes to handle card buying and selling, and money subtraction operations, were added.