Feat/uorchestrator (#13)
* Add orchestrator project

* Remove authentication and card services

This commit deletes all files related to the authentication and card services. These services include AuthController, AuthService, CardService, and associated helper classes and data transfer objects. These files were removed as part of refactoring and cleanup process to simplify the overall codebase.

* Add Lombok dependency and update UserAdd class

A dependency for the Lombok library has been added to the project. The 'UserAdd' class was updated to implement the 'JavaDelegate' interface, extending its functionality with a rollback mechanism for address update tasks. Also, Lombok annotations were added for cleaner and simpler code.

* Add AuthDTO class and modify UserAdd class

In this commit, a new AuthDTO class has been added to carry authentication data. The UserAdd class in the orchestrator module has been modified to include a mechanism for creating a new user via a REST call. The method now extracts user details from the delegateExecution, sends them to the user service, and handles the response.

* Correct all the Orchestrator

* Update orchestrator configuration and refactor register process

This commit updates several components in the orchestrator. It removes the application.properties file and introduces a bpm-plateform.xml file for Camunda configuration. The pom.xml dependencies are updated, shifting to Spring Boot 3.2.4 and adjusting related configurations, including compile and repackage settings. A new security configuration is introduced in the Gateway. Register process classes are refactored from "com.sp.registrer" to "com.sp.register".

* Update orchestrator configuration and refactor register process

This commit updates several components in the orchestrator. It removes the application.properties file and introduces a bpm-plateform.xml file for Camunda configuration. The pom.xml dependencies are updated, shifting to Spring Boot 3.2.4 and adjusting related configurations, including compile and repackage settings. A new security configuration is introduced in the Gateway. Register process classes are refactored from "com.sp.registrer" to "com.sp.register".

* First process work

* Refactor and enhance handling of card, money, and user operations

This commit refactors the handling of operations associated with cards, money, and user. It includes renaming of the `CardCheck` class to `CardBuyable` and deletion of the `bpm-plateform.xml` file. It also handles access failures to the platform more robustly by catching RestClientExceptions and throwing BpmnErrors in response. New classes to handle card buying and selling, and money subtraction operations, were added.